+++ Engine/Assets/Shaders/Object.fs.glsl
@@ -1,0 +1,7 @@
+#version 330 core
+
+out vec4 aColor;
+
+void main() {
+    aColor = vec4(1.0, 1.0, 1.0, 1.0);
+}