+++ Engine/Source/Core/Engine.cpp
@@ -4,6 +4,13 @@
 
 #include "Layer/LayerStack.h"
 
+#include "Renderer/Renderer.h"
+
+#include <string>
+#include <iostream>
+#include <fstream>
+#include <vector>
+
 namespace Core
 {
     static Engine::State state;
@@ -21,6 +28,11 @@ namespace Core
         WindowInformation.Height = 720;
         WindowInformation.Title = "Hello Core";
         state._Window = new Window(WindowInformation);
+
+        Renderer::InitializeRenderingContext();
+        Renderer::Init();
+
+        Renderer::Viewport(state._Window->GetInfo()->Width, state._Window->GetInfo()->Height);
     }
 
     void Engine::Init()
@@ -40,8 +52,12 @@ namespace Core
 
     void Engine::Render()
     {
+        Renderer::BeginFrame();
+        Renderer::Render();
         if (state._GApp)
             state._GApp->Render();
+        Renderer::EndFrame();
+        Renderer::RenderScreenImage();
     }
 
     void Engine::Shutdown()
@@ -57,6 +73,7 @@ namespace Core
         Logger::Shutdown();
         LayerStack::Shutdown();
         Input::Shutdown();
+        Renderer::Shutdown();
     }
 
     bool Engine::ShouldRun()
@@ -72,5 +89,26 @@ namespace Core
     void Engine::FeedApplication(Application *_App)
     {
         state._GApp = _App;
+    }
+
+    std::string Engine::ReadFileContent(const std::string &filename)
+    {
+        std::string vertResult;
+        std::ifstream inVert(filename, std::ios::in | std::ios::binary);
+        if (inVert)
+        {
+
+            inVert.seekg(0, std::ios::end);
+            vertResult.resize(inVert.tellg());
+            inVert.seekg(0, std::ios::beg);
+            inVert.read(&vertResult[0], vertResult.size());
+            inVert.close();
+        }
+        else
+        {
+            CE_CORE_ERROR("Unable to read file content:  %s", filename.c_str());
+            vertResult = "";
+        }
+        return vertResult;
     }
 }