+++ Engine/Source/Renderer/GPUScreen.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "Base.h"
+#include "Buffer/Buffer.h"
+#include "Buffer/FrameBuffer.h"
+#include "Buffer/VertexArray.h"
+#include "Shader/Shader.h"
+
+namespace Core
+{
+    struct GPUScreen
+    {
+        VertexArray *Array;
+        FrameBuffer *Buffer;
+        Shader *SShader;
+
+        void Setup();
+        void Begin();
+        void End();
+    };
+}